Refactor writing down cluster creds

+++ runhouse/resources/secrets/provider_secrets/ssh_secret.py
@@ -90,11 +90,11 @@ class SSHSecret(ProviderSecret):
 
         priv_key_path.parent.mkdir(parents=True, exist_ok=True)
         private_key = values.get("private_key")
-        if private_key is not None:
+        if private_key is not None and not priv_key_path.exists():
             priv_key_path.write_text(private_key)
             priv_key_path.chmod(0o600)
         public_key = values.get("public_key")
-        if public_key is not None:
+        if public_key is not None and not pub_key_path.exists():
             pub_key_path.write_text(public_key)
             pub_key_path.chmod(0o600)
